+++ src/test/java/org/primefaces/extensions/integrationtests/datepicker/DatePicker001IT.java
@@ -1,7 +1,5 @@
 package org.primefaces.extensions.integrationtests.datepicker;
 
-import org.apache.tomee.embedded.Configuration;
-import org.apache.tomee.embedded.Container;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Tag;
@@ -11,13 +9,11 @@ import org.primefaces.extensions.integrationtests.BaseIT;
 @Tag("DatePicker Tests")
 public class DatePicker001IT extends BaseIT {
     private static final String datePickerInputID = "$('[id$=datepicker1_input]')";
-    private static final String datePickerPanelID = "$('[id$=datepicker1_panel]')";
 
-    @Override
     @DisplayName("Test Input")
     public void performTest() {
         focus(datePickerInputID);
-        setValue(datePickerInputID,"02/19/1978");
+        setValue(datePickerInputID, "02/19/1978");
         blur(datePickerInputID);
 
         Assertions.assertTrue(doesExist("$(\"[id$=datepicker1_panel] .ui-datepicker-month:contains('February')\")"));
@@ -29,8 +25,9 @@ public class DatePicker001IT extends BaseIT {
         Assertions.assertEquals("02/19/1978", valueAfterClick);
     }
 
-    public String getLocation()
-    {
+    public String getLocation() {
         return "datepicker/datePicker001.xhtml";
     }
+
+
 }
